pyproject.toml
[tool.poetry]
name = "petroly-django"
version = "1.0.0"
description = "Petroly Initiative backend and API"
authors = ["ammar-faifi <a@ammarf.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "3.11.6"
Django = "^4.2.6"
django-q = "^1.3.9"
django-choices-field = "^2.2"
django-cors-headers = "^4.3.0"
gunicorn = "^21.2.0"
Pillow = "^10.0.1"
psycopg2 = "^2.9.9"
strawberry-django-auth = "^0.376.2"
strawberry-graphql = "^0.211.1"
strawberry-graphql-django = "^0.23.0"
whitenoise = "^6.6.0"
cloudinary = "^1.36.0"
python-telegram-bot = {extras = ["callback-data"], version = "^20.7"}
django-multiselectfield = "^0.1.12"
cryptography = "^41.0.7"
requests = "^2.31.0"
django-email-verification = "^0.3.3"
django-filter = "^23.3"
dj-database-url = "^2.0.0"
django-debug-toolbar = "^4.2.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
openai = "^1.6.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
mypy = "^1.4"
pylint-django = "^2.5.3"
flake8 = "^6.1.0"
django-stubs = "^4.2.3"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
